frontend/requirements.txt
streamlit
requests
minio
pandas
python-dotenv
